Make entry point finding reusable so we can easily find which blocks to skip in the old build
