+++ plugins/woocommerce/client/blocks/bin/webpack-config-interactivity-blocks-frontend.js
@@ -15,29 +15,9 @@ const RemoveFilesPlugin = require( './remove-files-webpack-plugin' );
  */
 const { sharedOptimizationConfig } = require( './webpack-shared-config' );
 const {
-	findInteractivityBlockAssets,
-} = require( './find-interactivity-block-assets' );
-
-const interactivityBlocks = findInteractivityBlockAssets(
-	path.resolve( __dirname, '../assets/js' ),
-	[ 'frontend.*s', 'style.scss' ]
-);
-
-const scriptModuleEntries = interactivityBlocks.reduce( ( acc, block ) => {
-	const frontendFile = block.assets.find( ( f ) => f.includes( 'frontend' ) );
-	if ( frontendFile ) {
-		acc[ block.blockName ] = frontendFile;
-	}
-	return acc;
-}, {} );
-
-const styleEntries = interactivityBlocks.reduce( ( acc, block ) => {
-	const styleFile = block.assets.find( ( f ) => f.includes( 'style' ) );
-	if ( styleFile ) {
-		acc[ `${ block.blockName }-style` ] = styleFile;
-	}
-	return acc;
-}, {} );
+	scriptModuleEntries,
+	styleEntries,
+} = require( './webpack-interactivity-entries' );
 
 const entries = {
 	// Blocks
